Move match integration tests to their own folder
